Move fieldID management to id package
FieldID is really a 'short id', as long term it will have a CID and the uint will be managed internally, just like collection short ids (previously called 'Root').

All short ids will be managed (and cached) here in this new sub-package, including the short CollectionID in a commit later in this PR.
